% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shinyrates_data.R
\name{format_shinyrates_data}
\alias{format_shinyrates_data}
\title{Format the shinyrates data}
\usage{
format_shinyrates_data(df, timestamp)
}
\arguments{
\item{df:}{a data frame to format}

\item{timestamp:}{the specific timestamp to use}
}
\value{
A data.frame with the following columns:
     \item{date_recorded}{ Date the data was pulled}
     \item{pokemon_ID}{National Pokédex Number}
     \item{pokemon_name}{Name of the Pokemon}
     \item{shiny_rate_frac}{Shiny rate as a fraction as pulled from shinyrates.com}
     \item{shiny_rate}{Decimal shiny rate calubluated from `shiny_rate_frac`}
     \item{sample_size}{Number of that species of pokemon observed in the last 24 hours}
}
\description{
Formats the data from shinyrates.com
}
\examples{
date <- Sys.Date()
df <- scrape_shinyrates_website()
format_shinyrates_data(df=df, timestamp=date)

}
