% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shinyrates_data.R
\name{get_shinyrates_data}
\alias{get_shinyrates_data}
\title{Get shiny rates data}
\usage{
get_shinyrates_data(page = "https://shinyrates.com/data/rate")
}
\arguments{
\item{page:}{the page to pull the shinyrates JSON file from (default is https://shinyrates.com/data/rate)}
}
\value{
A data.frame with the following columns:
     \item{date_recorded}{ Date the data was pulled}
     \item{pokemon_ID}{National Pokédex Number}
     \item{pokemon_name}{Name of the Pokemon}
     \item{shiny_rate_frac}{Shiny rate as a fraction as pulled from shinyrates.com}
     \item{shiny_rate}{Decimal shiny rate calubluated from `shiny_rate_frac`}
     \item{sample_size}{Number of that species of pokemon observed in the last 24 hours}
}
\description{
scrapes the data from shinyrates.com and formats it as a dataframe
}
\examples{
get_shinyrates_data()

}
