% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shinyrates_data.R
\name{scrape_shinyrates_website}
\alias{scrape_shinyrates_website}
\title{Scrapes shinyrates.com}
\usage{
scrape_shinyrates_website(page)
}
\arguments{
\item{page:}{the page to pull the shinyrates JSON file from (default is https://shinyrates.com/data/rate)}
}
\value{
A data.frame with the following columns:
     \item{pokemon_ID}{National Pokédex Number}
     \item{pokemon_name}{Name of the Pokemon}
     \item{shiny_rate_frac}{Shiny rate as a fraction as pulled from shinyrates.com}
     \item{sample_size}{Number of that species of pokemon observed in the last 24 hours}
}
\description{
scrapes the data from shinyrates.com and formats it as a dataframe
}
\examples{
scrape_shinyrates_website()

}
