% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shinyrates_data.R
\name{write_shinyrates_data}
\alias{write_shinyrates_data}
\title{Write the shinyrates data.}
\usage{
write_shinyrates_data(df, name = "shinyrates.csv")
}
\arguments{
\item{df:}{a data frame to write}

\item{name:}{the name of the file (with file type ie name.csv)}
}
\description{
writes the data from shinyrates.com
}
\examples{
date <- Sys.Date()
df <- scrape_shinyrates_website()
df <- format_shinyrates_data(df=df, timestamp=date)
write_shinyrates_data(df)

}
